Revert creation of academy.exchange.exception
This created way to many circular dependency errors
