+++ academy/exchange/cloud/server.py
@@ -51,6 +51,8 @@ from aiohttp.web import run_app
 from pydantic import TypeAdapter
 from pydantic import ValidationError
 
+from academy.exception import BadEntityIdError
+from academy.exception import MailboxTerminatedError
 from academy.exchange import MailboxStatus
 from academy.exchange.cloud.authenticate import Authenticator
 from academy.exchange.cloud.authenticate import get_authenticator
@@ -58,8 +60,6 @@ from academy.exchange.cloud.config import ExchangeAuthConfig
 from academy.exchange.cloud.config import ExchangeServingConfig
 from academy.exchange.cloud.exceptions import ForbiddenError
 from academy.exchange.cloud.exceptions import UnauthorizedError
-from academy.exchange.exception import BadEntityIdError
-from academy.exchange.exception import MailboxTerminatedError
 from academy.identifier import AgentId
 from academy.identifier import EntityId
 from academy.logging import init_logging
