+++ academy/exchange/hybrid.py
@@ -34,9 +34,9 @@ import redis.asyncio
 
 from academy.behavior import Behavior
 from academy.behavior import BehaviorT
+from academy.exception import BadEntityIdError
+from academy.exception import MailboxTerminatedError
 from academy.exchange import ExchangeFactory
-from academy.exchange.exception import BadEntityIdError
-from academy.exchange.exception import MailboxTerminatedError
 from academy.exchange.redis import _MailboxState
 from academy.exchange.redis import _RedisConnectionInfo
 from academy.exchange.transport import ExchangeTransportMixin
