+++ academy/exchange/local.py
@@ -19,9 +19,9 @@ from culsans import Queue
 
 from academy.behavior import Behavior
 from academy.behavior import BehaviorT
+from academy.exception import BadEntityIdError
+from academy.exception import MailboxTerminatedError
 from academy.exchange import ExchangeFactory
-from academy.exchange.exception import BadEntityIdError
-from academy.exchange.exception import MailboxTerminatedError
 from academy.exchange.transport import ExchangeTransportMixin
 from academy.exchange.transport import MailboxStatus
 from academy.identifier import AgentId
